pyproject.toml
[tool.poetry]
name = "geotracking"
version = "0.1.0"
description = "The Ultimate Django and Vue Template"
authors = ["Roger Camargo"]
license = "MIT"
readme = "README.md"
packages = [{include = "geotracking"}]

[tool.poetry.dependencies]
python = "3.11.x"
django = "^4.2.1"
asgiref = "^3.6.0"
sqlparse = "^0.4.4"
tzdata = "^2023.3"
django-extensions = "^3.2.1"
psycopg2 = "^2.9.6"
dj-database-url = "^2.0.0"
python-decouple = "^3.8"
django-ninja = "^1.1.0"
whitenoise = "^6.4.0"
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-django = "^4.5.2"
ipython = "^8.17.1"
mock = "^5.1.0"
django-cors-headers = "^4.3.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "geotracking.geotracking.settings"
pythonpath = [
  "./geotracking"
]
addopts = "-x -vv"
filterwarnings = [
    "ignore::pytest.PytestCacheWarning",
    "ignore::UserWarning",
    "ignore:No directory at",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
create = true
in-project = true